refactor(foundation): move MultiTenantConfig + RoutingStrategy (A14)
Extract the two pure-data multi_tenant types to
`mockforge_foundation::multi_tenant_types`:

- `MultiTenantConfig` (enabled flag, routing strategy, prefix, default
  workspace, port mappings, etc.)
- `RoutingStrategy` enum (Path/Port/Both)

Both are leaf data types with no behavioral dependencies on core.
`MultiTenantWorkspaceRegistry`, `WorkspaceRouter`, and `TenantWorkspace`
stay in core because they depend on `Workspace`, `RouteRegistry`, and
`CentralizedRequestLogger`.

Core's `multi_tenant::registry` re-exports from foundation for
backwards compat.

mockforge-http and mockforge-ui updated to import from foundation
directly. Drops another ~20 deprecation warnings from http (511 from
531) plus ~1 from ui. All 574 http + 1818 core tests pass; workspace
clippy clean.
